requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sentence-transformers>=2.3.1
faiss-cpu
python-multipart
jinja2
numpy>=1.24.0
pydantic>=2.6.0
